+++ www/apps/ui/src/examples/date-picker-range-time.tsx
@@ -1,0 +1,9 @@
+import { DatePicker } from "@medusajs/ui"
+
+export default function DatePickerRangeTime() {
+  return (
+    <div className="w-[250px]">
+      <DatePicker showTimePicker mode="range" />
+    </div>
+  )
+}